split & reorganize major defensive guide section
